+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/notifications/NotificationBuilder.kt
@@ -20,6 +20,5 @@ abstract class NotificationBuilder() {
         )
     }
 
-    abstract fun build(): Notification;
-
+    abstract fun build(): Notification
 }
