+++ webtrit_callkeep_android/android/src/main/kotlin/com/webtrit/callkeep/services/dispatchers/IncomingCallEventDispatcher.kt
@@ -1,8 +1,9 @@
 package com.webtrit.callkeep.services.dispatchers
 
 import android.content.Context
+import com.webtrit.callkeep.common.Log
 import com.webtrit.callkeep.models.CallMetadata
-import com.webtrit.callkeep.services.IncomingCallService
+import com.webtrit.callkeep.services.PushNotificationService
 import com.webtrit.callkeep.services.SignalingService
 
 interface IncomingCallServiceContract {
@@ -25,7 +26,7 @@ object IncomingCallAdapter : IncomingCallServiceContract {
         if (SignalingService.isRunning) {
             SignalingService.answerCall(context, metadata)
         } else {
-            IncomingCallService.answer(context, metadata)
+            PushNotificationService.answer(context, metadata)
         }
     }
 
@@ -33,12 +34,14 @@ object IncomingCallAdapter : IncomingCallServiceContract {
         if (SignalingService.isRunning) {
             SignalingService.endCall(context, metadata)
         } else {
-            IncomingCallService.hangup(context, metadata)
+            PushNotificationService.hangup(context, metadata)
         }
     }
 }
 
 object IncomingCallEventDispatcher {
+    const val TAG = "IncomingCallEventDispatcher"
+
     private fun getActiveService(): IncomingCallServiceContract {
         return if (SignalingService.isRunning) {
             ForegroundCallAdapter
@@ -48,10 +51,12 @@ object IncomingCallEventDispatcher {
     }
 
     fun answer(context: Context, metadata: CallMetadata) {
+        Log.d(TAG, "answer: $metadata")
         getActiveService().answerIncomingCall(context, metadata)
     }
 
     fun hungUp(context: Context, metadata: CallMetadata) {
+        Log.d(TAG, "hungUp: $metadata")
         getActiveService().declineIncomingCall(context, metadata)
     }
 }
